+++ packages/components/src/index.ts
@@ -11,3 +11,4 @@ export { MinisTag } from './components/tag/tag.js';
 export { MinisNavigation } from './components/navigation/navigation.js';
 export { MinisNavigationItem } from './components/navigation/navigation-item.js';
 export { MinisCardGrid } from './components/card-grid/card-grid.js';
+export { MinisContainer } from './components/container/container.js';